+++ tests/metadata_test.go
@@ -9,3 +9,13 @@ import (
 func TestMetadata(t *testing.T) {
 	tooling.LogVersion(t)
 }
+
+const (
+	GroupTrustlessGateway = "Trustless Gateway"
+	GroupPathGateway      = "Path Gateway"
+	GroupSubdomainGateway = "Subdomain Gateway"
+)
+
+const (
+	IPIP402 = "ipip-0402"
+)