+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `--version` flag shows the current version
 - Metadata logging used to associate tests with custom data like versions, specs identifiers, etc.
+- Test Group Metadata on Tests
+- IPIP Metadata on Tests and SugarTest
 
 ## [0.3.0] - 2023-07-31
 ### Added